calorie-app/ai-model/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow-cpu
